+++ CHANGELOG.rst
@@ -1,7 +1,15 @@
 Changelog
 =========
 
-0.0.5 (Current)
+0.0.6 (current)
 ---------------
+
+* "turma" in `/periodos/2016s2/oferecimentos/MC878/turma/a/matriculados`
+  changes to its plural form "turmas".
+* Parameters pre-processing (including sensitivity) are now declared
+  in the views themselves, through the attribute `route_parameters`.
+
+0.0.5
+-----
 
 * Started tracking.